extract test into spec file
