+++ gb-cpu/src/lib.rs
@@ -2,3 +2,5 @@ pub mod bus;
 pub mod cpu;
 pub mod interfaces;
 pub mod registers;
+#[cfg(test)]
+mod registers_test;
